Add files via upload
An online auction system that let the mobile users participate in an auction session. The data provided by the user; be it sellers or bidders, is stored in the database. The user can upload the products that he/she wishes to sell and can also bid on the items posted by the other users. After the bidding session is expired, the highest bidder is announced as a winner. It also has an option of signing up for the voice notifications in the event when a product is sold.